added tags to prepare files for relocation pt2
