+++ relecov_dashboard/utils/generic/generic_graphic_data.py
